+++ src/App.tsx
@@ -1,106 +1,24 @@
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useState } from 'react';
 import './App.scss';
-import { peopleFromServer } from './data/people';
-import debounce from 'lodash.debounce';
+import { Autocomplete } from './components/Autocomplete';
+import { peopleFromServer, Person } from './data/people';
 
-type Props = {
-  onSelected?: (person: Person) => void;
-};
-
-export const App: React.FC<Props> = ({ onSelected }) => {
-  const [query, setQuery] = useState('');
-  const [isFocused, setIsFocused] = useState(false);
+export const App: React.FC<Props> = () => {
   const [selectedPerson, setSelectedPerson] = useState<Person | null>(null);
-  const [appliedQuery, setAppliedQuery] = useState('');
-
-  const applyQuery = useCallback(debounce(setAppliedQuery, 1000), []);
-
-  const handleQueryChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setQuery(event.target.value);
-    setSelectedPerson(null);
-    applyQuery(event.target.value);
-  };
-
-  const filteredNames = useMemo(() => {
-    const q = appliedQuery.trim().toLowerCase();
-
-    if (q === '' && isFocused) {
-      return peopleFromServer;
-    }
-
-    if (q === '') {
-      return [];
-    }
-
-    return peopleFromServer.filter(p => p.name.toLowerCase().includes(q));
-  }, [appliedQuery, isFocused]);
 
   return (
     <div className="container">
       <main className="section is-flex is-flex-direction-column">
-        <h1 className="title" data-cy="title">
+        <h1 className="title" data-qa="title">
           {selectedPerson
             ? `${selectedPerson.name} (${selectedPerson.born} - ${selectedPerson.died})`
             : 'No selected person'}
         </h1>
-
-        <div className={`dropdown ${isFocused ? 'is-active' : ''}`}>
-          <div className="dropdown-trigger">
-            <input
-              type="text"
-              placeholder="Enter a part of the name"
-              className="input"
-              data-cy="search-input"
-              value={query}
-              onChange={handleQueryChange}
-              onFocus={() => setIsFocused(true)}
-              onBlur={() => setIsFocused(false)}
-            />
-          </div>
-
-          {isFocused && (
-            <div
-              className="dropdown-menu"
-              role="menu"
-              data-cy="suggestions-list"
-            >
-              <div className="dropdown-content">
-                {filteredNames.length > 0 ? (
-                  filteredNames.map(person => (
-                    <div
-                      key={person.slug}
-                      className="dropdown-item"
-                      data-cy="suggestion-item"
-                      onMouseDown={e => e.preventDefault()}
-                      onClick={() => {
-                        setQuery(person.name);
-                        setSelectedPerson(person);
-                        onSelected?.(person);
-                        setIsFocused(false);
-                      }}
-                    >
-                      <p className="has-text-link">{person.name}</p>
-                    </div>
-                  ))
-                ) : (
-                  <div
-                    className="
-                    notification
-                    is-danger
-                    is-light
-                    mt-3
-                    is-align-self-flex-start
-                    "
-                    role="alert"
-                    data-cy="no-suggestions-message"
-                  >
-                    <p className="has-text-danger">No matching suggestions</p>
-                  </div>
-                )}
-              </div>
-            </div>
-          )}
-        </div>
+        <Autocomplete
+          people={peopleFromServer}
+          onSelected={setSelectedPerson}
+          delay={300}
+        />
       </main>
     </div>
   );
